Mostly formatting, also peak detector refactoring and peak extractor workflow changes.

+++ src/be/hogent/tarsos/peak/LocalVolumeScore.java
@@ -6,10 +6,10 @@ import be.hogent.tarsos.util.histogram.Histogram;
 
 
 public class LocalVolumeScore implements PeakScore {
-	
+
 	private final long[] volumes;
 	private final int windowSize;
-	
+
 	public LocalVolumeScore(Histogram originalHistogram,int windowSize){
 		volumes = new long[originalHistogram.getNumberOfClasses()];
 		this.windowSize = windowSize;
@@ -21,22 +21,22 @@ public class LocalVolumeScore implements PeakScore {
 			volumes[0] = volumes[0] + originalHistogram.getCountForClass(after);
 		}
 		volumes[0] = volumes[0] + originalHistogram.getCountForClass(0);
-		
+
 		//from now iterate histogram and use the first volume to calculate
 		//the other volumes
 		for(int i=1;i<originalHistogram.getNumberOfClasses();i++){
 			int after = i + windowSize;
 			int before = i - windowSize-1;
-			volumes[i] = volumes[i-1] + 
-			originalHistogram.getCountForClass(after) - 
+			volumes[i] = volumes[i-1] +
+			originalHistogram.getCountForClass(after) -
 			originalHistogram.getCountForClass(before);
 		}
 	}
-	
+
 	public double getVolumeAt(int index){
 		return volumes[index];
 	}
-	
+
 	@Override
 	public double score(Histogram originalHistogram, int index, int windowSize) {
 		assert this.windowSize == windowSize;
@@ -45,15 +45,15 @@ public class LocalVolumeScore implements PeakScore {
 		double[] volumeRange = new double[windowSize * 2 + 1];
 		int volumeRangeIndex = 0;
 		for(int j = 0; j < windowSize; j++ ){
-			before--;			
+			before--;
 			after++;
-			volumeRange[volumeRangeIndex] = volumes[(int) ( (index + before + originalHistogram.getNumberOfClasses() ) % originalHistogram.getNumberOfClasses()) ];
+			volumeRange[volumeRangeIndex] = volumes[( (index + before + originalHistogram.getNumberOfClasses() ) % originalHistogram.getNumberOfClasses()) ];
 			volumeRangeIndex++;
-			volumeRange[volumeRangeIndex] = volumes[(int) ((index + after) % originalHistogram.getNumberOfClasses()) ];
+			volumeRange[volumeRangeIndex] = volumes[((index + after) % originalHistogram.getNumberOfClasses()) ];
 			volumeRangeIndex++;
 		}
 		volumeRange[volumeRangeIndex] = volumes[index];
-		
+
 		double mean = StatUtils.mean(volumeRange);
 		double standardDeviation = Math.pow(StatUtils.variance(volumeRange,mean),0.5);
 		return standardDeviation == 0.0 ? 0.0 : (volumes[index] - mean ) / standardDeviation;
